moved API to pause all triggers for a workflow to trigger resource
