+++ common/src/main/java/com/cognitree/kronos/model/TaskUpdate.java
@@ -21,7 +21,6 @@ import com.cognitree.kronos.model.Task.Status;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
