+++ scheduler/src/test/java/com.cognitree.kronos/scheduler/WorkflowSchedulerServiceTest.java
@@ -27,7 +27,6 @@ import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.quartz.CronExpression;
-import org.quartz.JobKey;
 import org.quartz.Scheduler;
 import org.quartz.SchedulerException;
